% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/initialize.R
\docType{class}
\name{AbundanceMatrix-class}
\alias{AbundanceMatrix-class}
\alias{.AbundanceMatrix}
\alias{AbundanceMatrix}
\title{Relative Frequency Matrix}
\usage{
AbundanceMatrix(data = 0, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL)
}
\arguments{
\item{data}{A data vector.}

\item{nrow}{An \code{\link{integer}} value giving the desired number of rows.}

\item{ncol}{An \code{\link{integer}} value giving the desired number of
columns.}

\item{byrow}{A \code{\link{logical}} scalar: should the matrix be filled by
rows? If \code{FALSE} (the default) the matrix is filled by columns.}

\item{dimnames}{A list of length 2 giving the row and column names
respectively. If \code{NULL} (the default) dimension names will be created.}
}
\description{
An S4 class to represent a relative frequency matrix (i.e. the fraction of
times a given datum occurs in a dataset).
}
\section{Slots}{

\describe{
\item{\code{totals}}{A \code{\link{numeric}} vector giving the absolute row sums.}
}}

\section{Access}{

 In the code snippets below, \code{x} is a \code{*Matrix} object.
 \describe{
  \item{\code{length(x)}}{Returns the length of \code{x}.}
  \item{\code{dim(x)}}{Returns the dimension of \code{x}.}
  \item{\code{nrow(x)}}{Returns the number of rows present in \code{x}.}
  \item{\code{ncol(x)}}{Returns the number of columns present in \code{x}.}
  \item{\code{dimnames(x)}, \code{dimnames(x) <- value}}{Retrieves or sets
  the row dimnames of \code{x} according to \code{value}.}
  \item{\code{rownames(x)}, \code{rownames(x) <- value}}{Retrieves or sets
  the row names of \code{x} according to \code{value}.}
  \item{\code{colnames(x)}, \code{colnames(x) <- value}}{Retrieves or sets
  the column names of \code{x} according to \code{value}.}
 }
}

\section{Subset}{

 In the code snippets below, \code{x} is a \code{*Matrix} object.
 \describe{
  \item{\code{x[i, j, ..., drop]}}{Extracts elements selected by subscripts
  \code{i} and \code{j}. Indices are \code{\link{numeric}} or
  \code{\link{character}} vectors or empty (missing) or \code{NULL}.
  Numeric values are coerced to \code{\link{integer}} as by
  \code{\link{as.integer}} (and hence truncated towards zero).
  Character vectors will be matched to the \code{\link{dimnames}} of the
  object. An empty index (a comma separated blank) indicates that all
  entries in that dimension are selected.
  Returns an object of the same class as \code{x}.}
  \item{\code{x[[i]]}}{Extracts a single element selected by subscript
  \code{i}.}
 }
}

\examples{
## Create an incidence (presence/absence) matrix
## Data will be coerced with as.logical()
A <- IncidenceMatrix(data = sample(0:1, 100, TRUE, c(1, 1/3)), nrow = 20)
## Create a count data matrix
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20)

## Access
dim(B) # Get the matrix dimensions
row(B) # Get the row indexes
col(B, as.factor = TRUE) # Get the column indexes
nrow(B) # Get the number of rows
ncol(B) # Get the number of columns
dimnames(B) # Get the dimension names
rownames(B) <- LETTERS[1:20] # Set the row names
rownames(B) # Get the rownames
colnames(B) <- letters[21:25] # Set the column names
colnames(B) # Get the column names

## Subset
B[[1, 1]] # Get the first value
B[, ] # Get all values
B[1, ] # Get the first row
B[, 1] # Get the first column
B[c("A", "B", "C"), ] # Get the first three rows
B[c("D", "E", "F"), 1, drop = FALSE] # Get 3 rows of the first column
}
\seealso{
\link{as_abundance}

Other matrix: 
\code{\link{CountMatrix-class}},
\code{\link{DataMatrix-class}},
\code{\link{GenericMatrix-class}},
\code{\link{IncidenceMatrix-class}},
\code{\link{OccurrenceMatrix-class}},
\code{\link{SimilarityMatrix-class}},
\code{\link{StratigraphicMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{matrix}
