% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkhe-deprecated.R
\name{arkhe-deprecated}
\alias{arkhe-deprecated}
\alias{confidence}
\alias{confidence-method}
\alias{confidence,numeric-method}
\title{Deprecated Functions in arkhe}
\usage{
confidence(object, ...)

\S4method{confidence}{numeric}(object, level = 0.95, type = c("student", "normal"))
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
