% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_constant}
\alias{assert_constant}
\alias{assert_decreasing}
\alias{assert_increasing}
\title{Check Numeric Trend}
\usage{
assert_constant(x, ...)

assert_decreasing(x, ...)

assert_increasing(x, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} object to be checked.}

\item{...}{Extra parameters to be passed to internal methods.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Numeric Trend
}
\seealso{
Other checking methods: 
\code{\link{assert_dim}()},
\code{\link{assert_empty}()},
\code{\link{assert_infinite}()},
\code{\link{assert_length}()},
\code{\link{assert_lower}()},
\code{\link{assert_missing}()},
\code{\link{assert_names}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_square}()},
\code{\link{assert_type}()},
\code{\link{assert_unique}()}
}
\author{
N. Frerebeau
}
\concept{checking methods}
