% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_data}
\alias{assert_data}
\alias{assert_missing}
\alias{assert_infinite}
\alias{assert_unique}
\title{Check Data}
\usage{
assert_missing(x)

assert_infinite(x)

assert_unique(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
\itemize{
\item \code{assert_missing()} and \code{assert_infinite()} check if an object contains any
missing (\code{NA}, \code{NaN}) or infinite (\code{Inf}) value.
\item \code{assert_unique()} checks if an object contains duplicated elements.
}
}
\seealso{
Other validation methods: 
\code{\link{assert_constant}()},
\code{\link{assert_length}()},
\code{\link{assert_lower}()},
\code{\link{assert_names}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_square}()},
\code{\link{assert_type}()},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
