% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_empty}
\alias{assert_empty}
\alias{assert_filled}
\title{Check Object Filling}
\usage{
assert_empty(x)

assert_filled(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Checks if an object is (not) empty.
}
\seealso{
Other checking methods: 
\code{\link{assert_constant}()},
\code{\link{assert_dim}()},
\code{\link{assert_infinite}()},
\code{\link{assert_length}()},
\code{\link{assert_lower}()},
\code{\link{assert_missing}()},
\code{\link{assert_names}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_square}()},
\code{\link{assert_type}()},
\code{\link{assert_unique}()}
}
\author{
N. Frerebeau
}
\concept{checking methods}
