% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_length}
\alias{assert_length}
\alias{assert_lengths}
\title{Check Object Length(s)}
\usage{
assert_length(
  x,
  expected,
  allow_empty = empty,
  allow_null = FALSE,
  empty = FALSE
)

assert_lengths(x, expected)
}
\arguments{
\item{x}{An object to be checked.}

\item{expected}{An appropriate expected value.}

\item{allow_empty}{A \code{\link{logical}} scalar: should \link[=is_empty]{empty} object be
ignored?}

\item{allow_null}{A \code{\link{logical}} scalar: should \code{NULL} object be ignored?}

\item{empty}{Deprecated.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Object Length(s)
}
\seealso{
Other checking methods: 
\code{\link{assert_constant}()},
\code{\link{assert_dim}()},
\code{\link{assert_empty}()},
\code{\link{assert_infinite}()},
\code{\link{assert_lower}()},
\code{\link{assert_missing}()},
\code{\link{assert_names}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_square}()},
\code{\link{assert_type}()},
\code{\link{assert_unique}()}
}
\author{
N. Frerebeau
}
\concept{checking methods}
