% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_lower}
\alias{assert_lower}
\alias{assert_greater}
\title{Check Numeric Relations}
\usage{
assert_lower(x, y, ...)

assert_greater(x, y, ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} object to be checked.}

\item{...}{Extra parameters to be passed to internal methods.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Numeric Relations
}
\seealso{
Other checking methods: 
\code{\link{assert_constant}()},
\code{\link{assert_dim}()},
\code{\link{assert_empty}()},
\code{\link{assert_infinite}()},
\code{\link{assert_length}()},
\code{\link{assert_missing}()},
\code{\link{assert_names}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_square}()},
\code{\link{assert_type}()},
\code{\link{assert_unique}()}
}
\author{
N. Frerebeau
}
\concept{checking methods}
