% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_names}
\alias{assert_names}
\alias{assert_dimnames}
\alias{assert_rownames}
\alias{assert_colnames}
\title{Check Object Names}
\usage{
assert_names(x, expected)

assert_dimnames(x, expected)

assert_rownames(x, expected)

assert_colnames(x, expected)
}
\arguments{
\item{x}{An object to be checked.}

\item{expected}{An appropriate expected value.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Object Names
}
\seealso{
Other validation methods: 
\code{\link{assert_constant}()},
\code{\link{assert_data}},
\code{\link{assert_length}()},
\code{\link{assert_lower}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_square}()},
\code{\link{assert_type}()},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
