% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/assign.R
\docType{methods}
\name{assign}
\alias{assign}
\alias{assign_colnames}
\alias{assign_colnames-method}
\alias{assign_rownames}
\alias{assign_rownames-method}
\alias{assign_rownames,data.frame-method}
\alias{assign_colnames,data.frame-method}
\title{Assign a specific row/column to the column/row names}
\usage{
assign_colnames(x, ...)

assign_rownames(x, ...)

\S4method{assign_rownames}{data.frame}(x, column, remove = TRUE)

\S4method{assign_colnames}{data.frame}(x, row, remove = TRUE)
}
\arguments{
\item{x}{A \code{\link{data.frame}}.}

\item{...}{Currently not used.}

\item{column}{A length-one \code{\link{numeric}} vector specifying the column number
that is to become the row names.}

\item{remove}{A \code{\link{logical}} scalar: should the specified row/column be removed
after making it the column/row names?}

\item{row}{A length-one \code{\link{numeric}} vector specifying the row number that is
to become the column names.}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Assign a specific row/column to the column/row names
}
\seealso{
Other data cleaning tools: 
\code{\link{count}()},
\code{\link{detect}()},
\code{\link{remove}()},
\code{\link{replace}()}
}
\author{
N. Frerebeau
}
\concept{data cleaning tools}
