% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap-method}
\alias{bootstrap,numeric-method}
\title{Bootstrap Estimation}
\usage{
bootstrap(x, ...)

\S4method{bootstrap}{numeric}(x, do, probs = c(0.05, 0.95), n = 1000, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{...}{Extra arguments passed to \code{do}.}

\item{do}{A \code{\link{function}} that takes \code{x} as an argument
and returns a single numeric value.}

\item{probs}{A \code{\link{numeric}} vector of probabilities with values in
\eqn{[0,1]} (see \code{\link[stats:quantile]{stats::quantile()}}).}

\item{n}{A non-negative \code{\link{integer}} giving the number of bootstrap
replications.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values be removed
from \code{x} before the sample statistics are computed?}
}
\value{
A \code{numeric} vector with the following elements:
\describe{
\item{\code{min}}{Minimum value.}
\item{\code{mean}}{Mean value.}
\item{\code{max}}{Maximum value.}
\item{\verb{Q*}}{Sample quantile to \code{*} probability.}
}
}
\description{
Bootstrap Estimation
}
\seealso{
Other statistics: 
\code{\link{jackknife}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
