% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CarbonDate-class}
\alias{CarbonDate-class}
\alias{.CarbonDate}
\title{Radiocarbon}
\description{
An S4 class to represent a radiocarbon date (BP, uncalibrated).
}
\seealso{
\linkS4class{NormalDate}

Other date: 
\code{\link{DateList-class}},
\code{\link{GenericDate-class}},
\code{\link{TypeDate-class}},
\code{\link{VirtualDate}}
}
\author{
N. Frerebeau
}
\concept{date}
\keyword{internal}
