% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-missing}
\alias{check-missing}
\alias{assert_missing}
\alias{assert_infinite}
\title{Check Missing Values}
\usage{
assert_missing(x)

assert_infinite(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Checks if an object contains any missing (\code{NA}, \code{NaN}) or infinite (\code{Inf})
value.
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
\keyword{internal}
