% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{check-numeric-comparison}
\alias{check-numeric-comparison}
\alias{assert_lower}
\alias{assert_greater}
\title{Check Numeric Relations}
\usage{
assert_lower(x, y, ...)

assert_greater(x, y, ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} object to be checked.}

\item{...}{Extra parameters to be passed to internal methods.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Numeric Relations
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-matrix}},
\code{\link{check-numeric-trend}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
