% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{check-numeric-trend}
\alias{check-numeric-trend}
\alias{assert_constant}
\alias{assert_decreasing}
\alias{assert_increasing}
\title{Check Numeric Trend}
\usage{
assert_constant(x, ...)

assert_decreasing(x, ...)

assert_increasing(x, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} object to be checked.}

\item{...}{Extra parameters to be passed to internal methods.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Numeric Trend
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-matrix}},
\code{\link{check-numeric-comparison}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{needs}()},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
