% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-type}
\alias{check-type}
\alias{check_type}
\alias{check_scalar}
\title{Check Data Types}
\usage{
check_type(x, expected)

check_scalar(x, expected)
}
\arguments{
\item{x}{An object to be checked.}

\item{expected}{A \code{\link{character}} string specifying the expected
type.}
}
\value{
Throw an error, if any.
}
\description{
Check Data Types
}
\seealso{
Other check: \code{\link{check-attribute}},
  \code{\link{check-graph}}, \code{\link{check-matrix}},
  \code{\link{check-missing}}, \code{\link{check-numeric}}
}
\author{
N. Frerebeau
}
\concept{check}
\keyword{internal}
