% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_class}
\alias{check_class}
\title{Class Diagnostic}
\usage{
check_class(object, conditions)
}
\arguments{
\item{object}{An object to which error messages are related.}

\item{conditions}{A \code{\link{list}} of condition messages.}
}
\value{
Throw an error if \code{conditions} is of non-zero length, invisibly returns
\code{TRUE} if not.
}
\description{
Class Diagnostic
}
\author{
N. Frerebeau
}
\keyword{internal}
