% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{circle}
\alias{circle}
\title{Draw a Circle}
\usage{
circle(x, y, radius, ..., n = 100)
}
\arguments{
\item{x, y}{A length-one \code{\link{numeric}} vector giving the coordinates of the
center of the circle.}

\item{radius}{A length-one \code{\link{numeric}} vector giving the radius of the
circle.}

\item{...}{Further parameters to be passed to \code{\link[graphics:polygon]{graphics::polygon()}}.}

\item{n}{A length-on \code{\link{integer}} vector specifying the number of vertices to
draw the circle.}
}
\description{
Draw a Circle
}
\note{
For internal use only.
}
\examples{
\dontrun{
plot(NA, xlim = c(-1, 1), ylim = c(-1, 1),
     axes = FALSE, ann = FALSE, asp = 1)
circle(0, 0, 0.5)
}
}
\seealso{
Other graphic tools: 
\code{\link{inch2line}()},
\code{\link{label_auto}()},
\code{\link{label_percent}()}
}
\author{
N. Frerebeau
}
\concept{graphic tools}
\keyword{internal}
