% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex-package.R
\docType{package}
\name{codex-package}
\alias{codex-package}
\alias{_PACKAGE}
\alias{codex}
\title{codex: Representation of Archaeological Data}
\description{
A collection of classes that represent
    archaeological data. This package provides a set of S4 classes that
    extend the basic matrix data type and a set of generic methods and
    functions. In addition, a few classes of general interest (e.g. that
    represent time and space information) are implemented.
}
\details{
\tabular{ll}{
 \strong{Package:} \tab codex \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.0.0.9000 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/xxx}{xxx}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://codex.archaeo.science}
  \item \url{http://github.com/nfrerebeau/codex}
  \item \url{https://cran.r-project.org/package=codex}
  \item Report bugs at \url{http://github.com/nfrerebeau/codex/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
 Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
