% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as_long}
\alias{as_long-method}
\alias{as_long,matrix-method}
\title{Coerce}
\usage{
as_long(from, ...)

\S4method{as_long}{matrix}(from, factor = FALSE, reverse = FALSE)
}
\arguments{
\item{from}{An object to be coerced.}

\item{...}{Currently not used.}

\item{factor}{A \code{\link{logical}} scalar: should character string be
coerced to \code{\link{factor}}? Default to \code{FALSE}, if \code{TRUE} the original ordering is
preserved.}

\item{reverse}{A \code{\link{logical}} scalar: should the order of factor levels be
reversed? Only used if \code{factor} is \code{TRUE}. Useful for plotting.}
}
\value{
A coerced object.
}
\description{
Coerces an object to a \verb{*Matrix} object.
}
\examples{
## Create a matrix
A <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Transform to long data.frame
head(as_long(A))
}
\author{
N. Frerebeau
}
\concept{classes}
