% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{concat}
\alias{concat}
\alias{\%+\%}
\title{Concatenate}
\usage{
x \%+\% y
}
\arguments{
\item{x, y}{A \code{\link{character}} vector.}
}
\value{
A \code{\link{character}} vector.
}
\description{
Concatenates character vectors.
}
\seealso{
Other utilities: 
\code{\link{null}}
}
\concept{utilities}
\keyword{utilities}
