% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{conditions}
\alias{conditions}
\alias{throw_error}
\alias{throw_warning}
\alias{catch_conditions}
\title{Conditions}
\usage{
throw_error(.subclass, message, call = NULL, ...)

throw_warning(.subclass, message, call = NULL, ...)

catch_conditions(expr)
}
\arguments{
\item{.subclass}{A \code{\link{character}} string specifying the class of
the message to be returned.}

\item{message}{A \code{\link{character}} string specifying the message to be
returned.}

\item{call}{The call.}

\item{...}{Extra arguments.}
}
\description{
\code{throw_error} stops execution of the current expression and executes an
error action.
}
\details{
\code{catch_conditions} handles unusual conditions.
}
\author{
N. Frerebeau
}
\concept{condition}
\keyword{error}
\keyword{internal}
