% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{confidence_bootstrap}
\alias{confidence_bootstrap}
\alias{confidence_bootstrap-method}
\alias{confidence_bootstrap,numeric-method}
\title{Nonparametric Bootstrap Confidence Interval}
\usage{
confidence_bootstrap(object, ...)

\S4method{confidence_bootstrap}{numeric}(
  object,
  level = 0.95,
  t0 = NULL,
  type = c("basic", "percentiles"),
  ...
)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector giving the bootstrap replicates of the
statistic of interest.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.
Must be a single number between \eqn{0} and \eqn{1}.}

\item{t0}{A length-one \code{\link{numeric}} vector giving the observed value of the
statistic of interest. Must be defined if \code{type} is "\code{basic}".}

\item{type}{A \code{\link{character}} string giving the type of confidence
interval to be returned. It must be one "\code{basic}" (the default) or
"\code{percentiles}". Any unambiguous substring can be given.}
}
\value{
A length-two \code{\link{numeric}} vector giving the lower and upper confidence
limits.
}
\description{
Computes equi-tailed two-sided nonparametric confidence interval.
}
\examples{
## Confidence interval for a mean
x <- seq(from = -4, to = 4, by = 0.01)
y <- dnorm(x)

mean(y)
confidence_mean(y, type = "student")
confidence_mean(y, type = "normal")

## Confidence interval for a propotion
confidence_binomial(118, n = 236)

x <- c(35, 74, 22, 69)
confidence_multinomial(x)
}
\references{
Davison, A. C. & Hinkley, D. V. (1997). \emph{Bootstrap Methods and Their
Application}. Cambridge Series on Statistical and Probabilistic Mathematics.
Cambridge: Cambridge University Press.
}
\seealso{
Other summary statistics: 
\code{\link{confidence_binomial}()},
\code{\link{confidence_mean}()},
\code{\link{confidence_multinomial}()},
\code{\link{interval_credible}()},
\code{\link{interval_hdr}()}
}
\author{
N. Frerebeau
}
\concept{summary statistics}
