% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{correlation}
\alias{correlation}
\alias{var,DataMatrix,missing-method}
\alias{var,DataMatrix-method}
\alias{cov,DataMatrix,missing-method}
\alias{cov,DataMatrix-method}
\alias{cor,DataMatrix,missing-method}
\alias{cor,DataMatrix-method}
\title{Correlation, Variance and Covariance}
\usage{
\S4method{var}{DataMatrix,missing}(x, na.rm = FALSE)

\S4method{cov}{DataMatrix,missing}(x, use = "everything", method = c("pearson", "kendall", "spearman"))

\S4method{cor}{DataMatrix,missing}(x, use = "everything", method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{A \linkS4class{DataMatrix} object.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values
(including \code{NaN}) be omitted from the calculations?}

\item{use}{An optional \code{\link{character}} string giving a method for
computing covariances in the presence of missing values (see
\code{\link[stats]{cor}}).}

\item{method}{A \code{\link{character}} string indicating which correlation
coefficient (or covariance) is to be computed (see
\code{\link[stats]{cor}}).}
}
\description{
Computes the variance, covariance or correlation of a \code{DataMatrix}.
}
\examples{
## Create count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Row and column means
rowMeans(A)
colMeans(A)

## Row and column sums
rowSums(A)
colSums(A)

## Variance, covariance and correlation matrix
var(A)
cov(A)
cor(A)
}
\seealso{
Other statistics: 
\code{\link{statistics}}
}
\author{
N. Frerebeau
}
\concept{statistics}
