% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ca.R
\docType{methods}
\name{correspondence}
\alias{correspondence}
\alias{ca}
\alias{ca-method}
\alias{get_coordinates}
\alias{get_coordinates-method}
\alias{get_contributions}
\alias{get_contributions-method}
\alias{get_distances}
\alias{get_distances-method}
\alias{get_eigenvalues}
\alias{get_eigenvalues-method}
\alias{get_inertia}
\alias{get_inertia-method}
\alias{ca,CountMatrix-method}
\alias{predict,CA-method}
\alias{get_coordinates,CA-method}
\alias{get_inertia,CA-method}
\alias{get_distances,CA-method}
\alias{get_contributions,CA-method}
\alias{get_eigenvalues,CA-method}
\title{Correspondence Analysis}
\usage{
ca(object, ...)

get_coordinates(object, ...)

get_contributions(object, ...)

get_distances(object, ...)

get_eigenvalues(object)

get_inertia(object, ...)

\S4method{ca}{CountMatrix}(object, n = 5, sup_rows = NULL, sup_columns = NULL)

\S4method{predict}{CA}(object, data, margin = 1)

\S4method{get_coordinates}{CA}(
  object,
  margin = 1,
  standard = FALSE,
  sup = TRUE,
  sup_name = ".sup"
)

\S4method{get_inertia}{CA}(object, margin = 1)

\S4method{get_distances}{CA}(object, margin = 1)

\S4method{get_contributions}{CA}(object, margin = 1)

\S4method{get_eigenvalues}{CA}(object)
}
\arguments{
\item{object}{A \linkS4class{CountMatrix} or a \linkS4class{CA} object.}

\item{...}{Currently not used.}

\item{n}{An \code{\link{integer}} value specifying the number of dimensions
to be kept in the results (default to \code{5}).}

\item{sup_rows}{An \code{\link{numeric}} vector specifying the indices of the
supplementary rows.}

\item{sup_columns}{An \code{\link{numeric}} vector specifying the indices of
the supplementary columns.}

\item{data}{An object of supplementary points coercible to a
\code{\link{matrix}} for which to compute principal coordinates.}

\item{margin}{A length-one \code{\link{numeric}} vector giving the subscript
which the data will be returned: \code{1} indicates rows (the default),
\code{2} indicates columns.}

\item{standard}{A \code{\link{logical}} scalar: should standard coordinates
be returned instead of principal coordinates (the default)?}

\item{sup}{A \code{\link{logical}} scalar: should supplementary points be
returned?}

\item{sup_name}{A \code{\link{character}} string specifying the name of the
column to create for supplementary points attribution (see below).}
}
\value{
\code{ca} returns a \linkS4class{CA} object.

 \code{predict} returns a \code{data.frame} of principal coordinates.

 \code{get_coordinates} returns a \code{data.frame} of coordinates. If
 \code{sup} is \code{TRUE}, an extra column (named after \code{sup_name}) is
 added specifying whether an observation is a supplementary point or not.

 \code{get_inertia} returns a \code{data.frame}.

 \code{get_distances} returns a \code{data.frame}.

 \code{get_contributions} returns a \code{data.frame}.

 \code{get_eigenvalues} returns a \code{data.frame} with the following
 columns: \code{eigenvalues}, \code{percentage} (percentage of variance) and
 \code{cumulative} (cumulative percentage of variance).
}
\description{
Computes a simple correspondence analysis based on the singular value
decomposition.
}
\examples{
## Create count data matrix
A <- CountMatrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

## Get row distances to centroid
get_distances(X, margin = 1)

## Get row inertias
get_inertia(X, margin = 1)

## Get row contributions
get_contributions(X, margin = 1)

## Get eigenvalues
get_eigenvalues(X)
}
\seealso{
Other multivariate analysis: 
\code{\link{CA-class}}
}
\author{
N. Frerebeau
}
\concept{multivariate analysis}
