% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DataMatrix-class}
\alias{DataMatrix-class}
\alias{IntegerMatrix-class}
\alias{.IntegerMatrix}
\alias{NumericMatrix-class}
\alias{.NumericMatrix}
\alias{LogicalMatrix-class}
\alias{.LogicalMatrix}
\title{Data Matrix}
\arguments{
\item{data}{A data vector.}

\item{nrow}{An \code{\link{integer}} value giving the desired number of rows.}

\item{ncol}{An \code{\link{integer}} value giving the desired number of
columns.}

\item{byrow}{A \code{\link{logical}} scalar: should the matrix be filled by
rows? If \code{FALSE} (the default) the matrix is filled by columns.}

\item{dimnames}{A list of length 2 giving the row and column names
respectively. If \code{NULL} (the default) dimension names will be created.}
}
\description{
A virtual S4 class to represent a data matrix.
}
\section{Slots}{

\describe{
\item{\code{values}}{An \code{\link{integer}}, a \code{\link{numeric}} or a
\code{\link{logical}} vector (see details).}
}}

\section{Access}{

 In the code snippets below, \code{x} is a \code{*Matrix} object.
 \describe{
  \item{\code{length(x)}}{Returns the length of \code{x}.}
  \item{\code{dim(x)}}{Returns the dimension of \code{x}.}
  \item{\code{nrow(x)}}{Returns the number of rows present in \code{x}.}
  \item{\code{ncol(x)}}{Returns the number of columns present in \code{x}.}
  \item{\code{dimnames(x)}, \code{dimnames(x) <- value}}{Retrieves or sets
  the row dimnames of \code{x} according to \code{value}.}
  \item{\code{rownames(x)}, \code{rownames(x) <- value}}{Retrieves or sets
  the row names of \code{x} according to \code{value}.}
  \item{\code{colnames(x)}, \code{colnames(x) <- value}}{Retrieves or sets
  the column names of \code{x} according to \code{value}.}
 }
}

\section{Subset}{

 In the code snippets below, \code{x} is a \code{*Matrix} object.
 \describe{
  \item{\code{x[i, j, ..., drop]}}{Extracts elements selected by subscripts
  \code{i} and \code{j}. Indices are \code{\link{numeric}} or
  \code{\link{character}} vectors or empty (missing) or \code{NULL}.
  Numeric values are coerced to \code{\link{integer}} as by
  \code{\link{as.integer}} (and hence truncated towards zero).
  Character vectors will be matched to the \code{\link{dimnames}} of the
  object. An empty index (a comma separated blank) indicates that all
  entries in that dimension are selected.
  Returns an object of the same class as \code{x}.}
  \item{\code{x[[i]]}}{Extracts a single element selected by subscript
  \code{i}.}
 }
}

\seealso{
Other matrix: 
\code{\link{AbundanceMatrix-class}},
\code{\link{CountMatrix-class}},
\code{\link{GenericMatrix-class}},
\code{\link{IncidenceMatrix-class}},
\code{\link{MatrixSummary-class}},
\code{\link{OccurrenceMatrix-class}},
\code{\link{SimilarityMatrix-class}},
\code{\link{StratigraphicMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{matrix}
\keyword{internal}
