% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DataMatrix}
\alias{DataMatrix}
\alias{IntegerMatrix-class}
\alias{.IntegerMatrix}
\alias{NumericMatrix-class}
\alias{.NumericMatrix}
\alias{LogicalMatrix-class}
\alias{.LogicalMatrix}
\title{Data Matrix}
\description{
S4 classes that represent a \eqn{m \times p}{m x p} matrix.
}
\note{
These classes inherit from \code{\link{matrix}}.
}
\seealso{
Other classes: 
\code{\link{CompositionMatrix-class}},
\code{\link{CountMatrix-class}},
\code{\link{IncidenceMatrix-class}},
\code{\link{OccurrenceMatrix-class}},
\code{\link{StratigraphicMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
