% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DateList-class}
\alias{DateList-class}
\alias{.DateList}
\title{Date}
\description{
An S4 class to represent a set of date.
}
\seealso{
Other date: 
\code{\link{CarbonDate-class}},
\code{\link{GenericDate-class}},
\code{\link{TypeDate-class}},
\code{\link{VirtualDate}}
}
\author{
N. Frerebeau
}
\concept{date}
\keyword{internal}
