% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecate.R
\docType{methods}
\name{arkhe-deprecated}
\alias{arkhe-deprecated}
\alias{as_frequency}
\alias{as_frequency-method}
\alias{as_frequency,ANY-method}
\title{Deprecated Methods}
\usage{
as_frequency(from)

\S4method{as_frequency}{ANY}(from)
}
\arguments{
\item{from}{A numeric \code{\link{matrix}} or \code{\link{data.frame}} to be
coerced.}
}
\description{
Deprecated Methods
}
\keyword{internal}
