% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GenericDate-class}
\alias{GenericDate-class}
\title{Date}
\description{
A virtual S4 class to represent a date. This is the mother class of all
date objects.
}
\section{Slots}{

\describe{
\item{\code{code}}{A \code{\link{character}} string specifying the unique
laboratory code of the date.}
}}

\seealso{
Other date: 
\code{\link{CarbonDate-class}},
\code{\link{DateList-class}},
\code{\link{TypeDate-class}},
\code{\link{VirtualDate}}
}
\author{
N. Frerebeau
}
\concept{date}
\keyword{internal}
