% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{inch2line}
\alias{inch2line}
\title{Plotting Dimensions of Character Strings}
\usage{
inch2line(x, ...)
}
\arguments{
\item{x}{A \code{\link{character}} vector of string whose length is to be calculated.}

\item{...}{Further parameter to be passed to \code{\link[graphics:strwidth]{graphics::strwidth()}}\verb{, such as }cex`.}
}
\value{
A \code{\link{numeric}} vector (maximum string width in units of margin lines).
}
\description{
Convert string length in inch to number of (margin) lines.
}
\note{
For internal use only.
}
\seealso{
Other graphic tools: 
\code{\link{circle}()},
\code{\link{label_auto}()},
\code{\link{label_percent}()}
}
\concept{graphic tools}
\keyword{internal}
