% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_scalar}
\alias{is_scalar}
\alias{is_scalar_list}
\alias{is_scalar_atomic}
\alias{is_scalar_vector}
\alias{is_scalar_numeric}
\alias{is_scalar_integer}
\alias{is_scalar_double}
\alias{is_scalar_character}
\alias{is_scalar_logical}
\title{Scalar Type Predicates}
\usage{
is_scalar_list(x)

is_scalar_atomic(x)

is_scalar_vector(x)

is_scalar_numeric(x)

is_scalar_integer(x)

is_scalar_double(x)

is_scalar_character(x)

is_scalar_logical(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Scalar Type Predicates
}
\seealso{
Other predicates: 
\code{\link{predicate-attributes}},
\code{\link{predicate-data}},
\code{\link{predicate-matrix}},
\code{\link{predicate-names}},
\code{\link{predicate-numeric}},
\code{\link{predicate-trend}},
\code{\link{predicate-type}}
}
\concept{predicates}
