% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{label_percent}
\alias{label_percent}
\title{Label Percentages}
\usage{
label_percent(x, digits = NULL, trim = FALSE)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{digits}{An \code{\link{integer}} indicating the number of decimal places.
If \code{NULL} (the default), breaks will have the minimum number of digits
needed to show the difference between adjacent values.}

\item{trim}{A \code{\link{logical}} scalar. If \code{FALSE} (the default), values are
right-justified to a common width (see \code{\link[base:format]{base::format()}}).}
}
\value{
A \code{\link{character}} vector.
}
\description{
Label Percentages
}
\keyword{internal}
