% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mathematics.R
\docType{methods}
\name{math_gcd}
\alias{math_gcd}
\alias{math_gcd-method}
\alias{math_gcd,numeric,numeric-method}
\title{Greatest Common Divisor}
\usage{
math_gcd(x, y)

\S4method{math_gcd}{numeric,numeric}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Computes the greatest common divisor (GCD) of two integer using the Euclidean
algorithm.
}
\seealso{
Other mathematic functions: 
\code{\link{math_lcm}()}
}
\author{
N. Frerebeau
}
\concept{mathematic functions}
