% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mathematics.R
\docType{methods}
\name{math_lcm}
\alias{math_lcm}
\alias{math_lcm-method}
\alias{math_lcm,numeric,numeric-method}
\title{Least Common Multiple}
\usage{
math_lcm(x, y)

\S4method{math_lcm}{numeric,numeric}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Computes the lowest common multiple of the denominators of a set of
fractions.
}
\seealso{
Other mathematic functions: 
\code{\link{math_gcd}()}
}
\author{
N. Frerebeau
}
\concept{mathematic functions}
