% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MatrixSummary-class}
\alias{MatrixSummary-class}
\alias{.MatrixSummary}
\title{Matrix Summary}
\description{
Matrix Summary
}
\seealso{
Other matrix: 
\code{\link{AbundanceMatrix-class}},
\code{\link{CountMatrix-class}},
\code{\link{DataMatrix-class}},
\code{\link{IncidenceMatrix-class}},
\code{\link{OccurrenceMatrix-class}},
\code{\link{SimilarityMatrix-class}},
\code{\link{StratigraphicMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{matrix}
\keyword{internal}
