% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{get_id}
\alias{get_id-method}
\alias{set_id<-}
\alias{set_id-method}
\alias{get_method}
\alias{get_method-method}
\alias{set_method<-}
\alias{set_method-method}
\alias{get_units}
\alias{get_units-method}
\alias{get_totals}
\alias{get_totals-method}
\alias{set_totals<-}
\alias{set_totals-method}
\alias{get_id,GenericMatrix-method}
\alias{get_totals,AbundanceMatrix-method}
\alias{get_method,SimilarityMatrix-method}
\alias{set_id<-,GenericMatrix-method}
\alias{set_id,GenericMatrix-method}
\alias{set_totals<-,AbundanceMatrix-method}
\alias{set_totals,AbundanceMatrix-method}
\alias{set_method<-,SimilarityMatrix-method}
\alias{set_method,SimilarityMatrix-method}
\title{Get or Set Parts of an Object}
\usage{
get_id(x)

set_id(x) <- value

get_method(x)

set_method(x) <- value

get_units(x)

get_totals(x)

set_totals(x) <- value

\S4method{get_id}{GenericMatrix}(x)

\S4method{get_totals}{AbundanceMatrix}(x)

\S4method{get_method}{SimilarityMatrix}(x)

\S4method{set_id}{GenericMatrix}(x) <- value

\S4method{set_totals}{AbundanceMatrix}(x) <- value

\S4method{set_method}{SimilarityMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{object}.}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\examples{
## Create an incidence (presence/absence) matrix
## Data will be coerced with as.logical()
A <- IncidenceMatrix(data = sample(0:1, 100, TRUE, c(1, 1/3)), nrow = 20)
## Create a count data matrix
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20)

## Access
dim(B) # Get the matrix dimensions
row(B) # Get the row indexes
col(B, as.factor = TRUE) # Get the column indexes
nrow(B) # Get the number of rows
ncol(B) # Get the number of columns
dimnames(B) # Get the dimension names
rownames(B) <- LETTERS[1:20] # Set the row names
rownames(B) # Get the rownames
colnames(B) <- letters[21:25] # Set the column names
colnames(B) # Get the column names

## Subset
B[[1, 1]] # Get the first value
B[, ] # Get all values
B[1, ] # Get the first row
B[, 1] # Get the first column
B[c("A", "B", "C"), ] # Get the first three rows
B[c("D", "E", "F"), 1, drop = FALSE] # Get 3 rows of the first column
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
