% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{needs}
\alias{needs}
\title{Check the Availability of a Package}
\usage{
needs(x, ask = TRUE)
}
\arguments{
\item{x}{A \code{\link{character}} vector naming the packages to check.}

\item{ask}{A \code{\link{logical}} scalar: should the user be asked to select packages
before they are downloaded and installed?}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Check the Availability of a Package
}
\details{
\code{needs()} is designed for use inside other functions in your own package
to check for the availability of a suggested package.

If the required packages are not available and \R is running interactively,
the user will be asked to install the packages.
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-matrix}},
\code{\link{check-numeric-comparison}},
\code{\link{check-numeric-trend}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
