% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{null}
\alias{null}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{An object.}
}
\value{
If \code{x} is \code{NULL}, returns \code{y}; otherwise returns \code{x}.
}
\description{
Replaces \code{NULL} with a default value.
}
\seealso{
Other utilities: 
\code{\link{concat}}
}
\concept{utilities}
\keyword{utilities}
