% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_color}
\alias{palette_color}
\alias{palette_color_continuous}
\alias{palette_color_discrete}
\title{Color Mapping}
\usage{
palette_color_continuous(x, palette = NULL, domain = NULL, na = "#DDDDDD", ...)

palette_color_discrete(
  x,
  palette = NULL,
  domain = NULL,
  ordered = FALSE,
  na = "#DDDDDD",
  ...
)
}
\arguments{
\item{x}{A vector of values.}

\item{palette}{A vector of colors or a color \code{function} that values will be
mapped to (see \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}}).}

\item{domain}{A \code{\link{numeric}} range or categorical data (according to \code{x}).}

\item{na}{The color to return for \code{NA} values.}

\item{...}{Currently not used.}

\item{ordered}{A \code{\link{logical}} scalar: should the levels be treated as already
in the correct order?}
}
\value{
A \code{\link{character}} vector of colors.
}
\description{
Color Mapping
}
\seealso{
Other palettes: 
\code{\link{palette_shape}()},
\code{\link{palette_size}()}
}
\concept{palettes}
