% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_shape}
\alias{palette_shape}
\title{Symbol Mapping}
\usage{
palette_shape(x, palette = NULL, domain = NULL, ordered = FALSE, ...)
}
\arguments{
\item{x}{A vector of categorical values.}

\item{palette}{A vector of symbols.}

\item{domain}{A vector of categorical data.}

\item{ordered}{A \code{\link{logical}} scalar: should the levels be treated as already
in the correct order?}

\item{...}{Currently not used.}
}
\value{
A vector of symbols.
}
\description{
Symbol Mapping
}
\seealso{
Other palettes: 
\code{\link{palette_color}},
\code{\link{palette_size}()}
}
\concept{palettes}
