% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-attributes}
\alias{predicate-attributes}
\alias{has_length}
\alias{is_empty}
\title{Attributes Predicates}
\usage{
has_length(x, n = NULL)

is_empty(x)
}
\arguments{
\item{x}{A \code{\link{vector}} to be tested.}

\item{n}{A length-one \code{\link{numeric}} vector specifying the length to test \code{x}
with. If \code{NULL}, returns \code{TRUE} if \code{x} has length greater than zero, and
\code{FALSE} otherwise.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
\itemize{
\item \code{has_length()} checks how long is an object.
\item \code{is_empty()} checks is an object is empty (any zero-length dimensions).
}
}
\seealso{
Other predicates: 
\code{\link{is_scalar}},
\code{\link{predicate-data}},
\code{\link{predicate-matrix}},
\code{\link{predicate-names}},
\code{\link{predicate-numeric}},
\code{\link{predicate-trend}},
\code{\link{predicate-type}}
}
\concept{predicates}
