% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-matrix}
\alias{predicate-matrix}
\alias{is_square}
\alias{is_symmetric}
\title{Matrix Predicates}
\usage{
is_square(x)

is_symmetric(x)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be tested.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
\itemize{
\item \code{is_square()} checks if a matrix is square.
\item \code{is_symmetric()} checks if a matrix is symmetric.
}
}
\seealso{
Other predicates: 
\code{\link{is_scalar}},
\code{\link{predicate-numeric}},
\code{\link{predicate-trend}},
\code{\link{predicate-type}},
\code{\link{predicate-utils}}
}
\concept{predicates}
