% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-names}
\alias{predicate-names}
\alias{has_names}
\alias{has_rownames}
\alias{has_colnames}
\title{Names Predicates}
\usage{
has_names(x, names = NULL)

has_rownames(x, names = NULL)

has_colnames(x, names = NULL)
}
\arguments{
\item{x}{A \code{\link{vector}} to be tested.}

\item{names}{A \code{\link{character}} vector specifying the names to test \code{x}
with. If \code{NULL}, returns \code{TRUE} if \code{x} has names, and \code{FALSE} otherwise.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Checks if an object is named.
}
\seealso{
Other predicates: 
\code{\link{is_scalar}},
\code{\link{predicate-attributes}},
\code{\link{predicate-data}},
\code{\link{predicate-matrix}},
\code{\link{predicate-numeric}},
\code{\link{predicate-trend}},
\code{\link{predicate-type}}
}
\concept{predicates}
