% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-type}
\alias{predicate-type}
\alias{is_list}
\alias{is_atomic}
\alias{is_vector}
\alias{is_numeric}
\alias{is_integer}
\alias{is_double}
\alias{is_character}
\alias{is_logical}
\alias{is_error}
\alias{is_warning}
\alias{is_message}
\title{Type Predicates}
\usage{
is_list(x)

is_atomic(x)

is_vector(x)

is_numeric(x)

is_integer(x)

is_double(x)

is_character(x)

is_logical(x)

is_error(x)

is_warning(x)

is_message(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Type Predicates
}
\seealso{
Other predicates: 
\code{\link{is_scalar}},
\code{\link{predicate-attributes}},
\code{\link{predicate-data}},
\code{\link{predicate-matrix}},
\code{\link{predicate-names}},
\code{\link{predicate-numeric}},
\code{\link{predicate-trend}}
}
\concept{predicates}
