% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/remove.R
\docType{methods}
\name{remove_whitespace}
\alias{remove_whitespace}
\alias{remove_whitespace-method}
\alias{remove_whitespace,data.frame-method}
\alias{remove_whitespace,matrix-method}
\title{Remove Leading/Trailing Whitespace}
\usage{
remove_whitespace(x, ...)

\S4method{remove_whitespace}{data.frame}(x, which = c("both", "left", "right"))

\S4method{remove_whitespace}{matrix}(x, which = c("both", "left", "right"))
}
\arguments{
\item{x}{An \R object (should be a \code{\link{matrix}} or a \code{\link{data.frame}}).}

\item{...}{Currently not used.}

\item{which}{A \code{\link{character}} string specifying whether to remove \code{both}
leading and trailing whitespace (default), or only leading ("\code{left}") or
trailing ("\code{right}").}
}
\description{
Remove Leading/Trailing Whitespace
}
\examples{
x <- data.frame(
  A = c("  Both. ", "  Left.", "Right.  "),
  B = 1:3
)

remove_whitespace(x, which = "both")
remove_whitespace(x, which = "left")
remove_whitespace(x, which = "right")
}
\seealso{
\code{\link[=trimws]{trimws()}}

Other data cleaning tools: 
\code{\link{empty}},
\code{\link{infinite}},
\code{\link{missing}},
\code{\link{remove_constant}()},
\code{\link{zero}}
}
\author{
N. Frerebeau
}
\concept{data cleaning tools}
