% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_range}
\alias{scale_range}
\title{Rescale Continuous Vector}
\usage{
scale_range(x, to = c(0, 1), from = range(x, finite = TRUE))
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{to}{A length-two \code{\link{numeric}} vector specifying the output range.}

\item{from}{A length-two \code{\link{numeric}} vector specifying the input range.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
Rescales continuous vector to have specified minimum and maximum.
}
\note{
For internal use only.
}
\concept{scales}
