% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,DataMatrix-method}
\title{Matrix Summaries}
\usage{
\S4method{summary}{DataMatrix}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{DataMatrix} object.}

\item{...}{Further arguments to be passed to \code{\link[base]{summary}}.}
}
\value{
A \linkS4class{MatrixSummary} object.
}
\description{
Matrix Summaries
}
\examples{
## Create a count data matrix
X <- CountMatrix(data = sample(1:10, 45, TRUE), nrow = 15)
summary(X)

## Summary by groups
set_groups(X) <- rep(c("A", "B", "C"), each = 5)
summary(X)
}
\author{
N. Frerebeau
}
\concept{summary}
