% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/chronology.R
\docType{methods}
\name{time}
\alias{time}
\alias{get_dates}
\alias{get_dates-method}
\alias{set_dates<-}
\alias{set_dates-method}
\alias{get_dates,AbundanceMatrix-method}
\alias{set_dates<-,AbundanceMatrix-method}
\alias{set_dates,AbundanceMatrix-method}
\title{Chronological Information}
\usage{
get_dates(object)

set_dates(object) <- value

\S4method{get_dates}{AbundanceMatrix}(object)

\S4method{set_dates}{AbundanceMatrix}(object) <- value
}
\arguments{
\item{object}{A \eqn{m \times p}{m x p} matrix of count data (typically of
class \linkS4class{CountMatrix}).}

\item{value}{A possible value for the element(s) of \code{object} (see
below).}
}
\description{
\code{set_dates} and \code{get_dates} allow to get or set the dates of an
object.
}
\details{
An attempt is made to interpret the argument \code{value} in a suitable way.
 \emph{Note} that errors are assumed to be given at \code{1} sigma.

 If \code{value} is a:
 \describe{
  \item{\code{character} vector}{it is assumed to contain Roman numerals.}
  \item{\code{numeric} or \code{integer} \code{vector}}{these values are
  assumed to represent dates without known errors.}
  \item{\code{list}}{containing components "\code{value}" and "\code{error}",
  these are used to define dates and corresponding errors.}
  \item{\code{matrix} or \code{data.frame} with two or more columns}{the
  first is assumed to contain the dates and the second the error values.
  \emph{Note} that if \code{value} has columns named "\code{value}" and
  "\code{error}", these columns will be used.}
 }
}
\author{
N. Frerebeau
}
\concept{chronology}
