% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TypeDate-class}
\alias{TypeDate-class}
\alias{.TypeDate}
\title{Typology}
\description{
An S4 class to represent a time span from typology (AD).
}
\seealso{
\linkS4class{UniformDate}

Other date: 
\code{\link{CarbonDate-class}},
\code{\link{DateList-class}},
\code{\link{GenericDate-class}},
\code{\link{VirtualDate}}
}
\author{
N. Frerebeau
}
\concept{date}
\keyword{internal}
