% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\alias{generate_uuid}
\alias{is_uuid}
\title{UUID}
\usage{
generate_uuid(seed = NULL)

is_uuid(x)
}
\arguments{
\item{seed}{A single \code{\link{integer}} specifying the seeds.
If \code{NULL} (the default) the seed will be re-initialized.}

\item{x}{A \code{\link{character}} string (UUID).}
}
\value{
\code{generate_uuid} returns a 36 characters long \code{\link{character}}
 string.

 \code{is_uuid} returns a \code{\link{logical}} scalar.
}
\description{
\code{generate_uuid} generates a universally unique identifier (UUID Version
4 and Variant 1).
}
\details{
\code{is_uuid} checks if a string is a canonically formatted UUID that is
Version 1 through 5 and is the appropriate Variant as per RFC4122.


 As it relies on R's internal random number generators and so will suffer
 from the use of \code{\link{set.seed}} in a session, the seed is
 re-initialized during execution (unless \code{seed} is not \code{NULL}).
 To prevent any side effects, the random number generator (RNG) state is
 saved and restored when the function exits.
}
\seealso{
\link{set.seed}
}
\author{
N. Frerebeau
}
\keyword{internal}
