% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{VirtualDate}
\alias{VirtualDate}
\alias{NormalDate-class}
\alias{UniformDate-class}
\title{Date Distribution}
\description{
A virtual S4 class to represent a data matrix.
}
\section{Slots}{

\describe{
\item{\code{mean,sd}}{An \code{\link{integer}} value giving the mean and standard
deviation of the distribution.}

\item{\code{min,max}}{An \code{\link{integer}} value giving the lower and upper
limits of the distribution.}
}}

\seealso{
Other date: 
\code{\link{CarbonDate-class}},
\code{\link{DateList-class}},
\code{\link{GenericDate-class}},
\code{\link{TypeDate-class}}
}
\author{
N. Frerebeau
}
\concept{date}
\keyword{internal}
