% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AbundanceMatrix-class}
\alias{AbundanceMatrix-class}
\alias{.AbundanceMatrix}
\title{Archaeological Data}
\description{
A virtual S4 class to represent archaeological data.
}
\section{Slots}{

\describe{
\item{\code{samples}}{A \code{\link{character}} vector.}

\item{\code{groups}}{A \code{\link{character}} vector.}

\item{\code{totals}}{A \code{\link{numeric}} vector giving the absolute row sums.}

\item{\code{dates_from}}{An \code{\link{integer}} vector.}

\item{\code{dates_to}}{An \code{\link{integer}} vector.}
}}

\section{Get and set}{

In the code snippets below, \code{x} is an \code{AbundanceMatrix} object.
\describe{
\item{\code{get_samples(x)} and \code{get_samples(x) <- value}}{Get or set
the sample names of \code{x}.}
\item{\code{get_groups(x)} and \code{set_groups(x) <- value}}{Get or set
the groups of \code{x}.}
\item{\code{get_dates(x)} and \code{set_dates(x) <- value}}{Get or set
the dates of \code{x}. \code{value} can be a \code{\link{list}} with components \code{from} (TPQ)
and \code{to} (TAQ) or a \code{\link{numeric}} vector (point estimate). The elements of
\code{value} are coerced to \code{\link{integer}} with \code{\link[=as.integer]{as.integer()}} (hence truncated
towards zero).}
}
}

\author{
N. Frerebeau
}
\keyword{internal}
