% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkhe-deprecated.R
\name{arkhe-deprecated}
\alias{arkhe-deprecated}
\alias{needs}
\alias{assert_dimensions}
\alias{compact_cols}
\alias{discard_cols}
\alias{keep_cols}
\title{Deprecated Functions in arkhe}
\usage{
needs(...)

assert_dimensions(...)

compact_cols(...)

discard_cols(...)

keep_cols(...)
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
