% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkhe-package.R
\docType{package}
\name{arkhe-package}
\alias{arkhe-package}
\alias{arkhe}
\title{arkhe: Tools for Cleaning Rectangular Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A dependency-free collection of simple functions for cleaning rectangular data. This package allows to detect, count and replace values or discard rows/columns using a predicate function. In addition, it provides tools to check conditions and return informative error messages.
}
\details{
\tabular{ll}{
\strong{Version} \tab 1.11.0 \cr
\strong{License} \tab GPL-3 \cr
\strong{CRAN DOI} \tab \doi{10.32614/CRAN.package.arkhe} \cr
\strong{Zenodo DOI} \tab \doi{10.5281/zenodo.3526659} \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\section{Package options}{

\pkg{arkhe} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{arkhe.verbose}: a \code{\link{logical}} scalar. Should \R report extra information
on progress? Defaults to \code{\link[=interactive]{interactive()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/arkhe}
  \item \url{https://packages.tesselle.org/arkhe/}
  \item Report bugs at \url{https://codeberg.org/tesselle/arkhe/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

Other contributors:
\itemize{
  \item Brice Lebrun (\href{https://orcid.org/0000-0001-7503-8685}{ORCID}) (Logo designer) [artist]
  \item Université Bordeaux Montaigne (\href{https://ror.org/03pbgwk21}{ROR}) [funder]
  \item CNRS (\href{https://ror.org/02feahw73}{ROR}) [funder]
}

}
\keyword{internal}
