% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkhe-package.R
\docType{package}
\name{arkhe-package}
\alias{arkhe-package}
\alias{_PACKAGE}
\alias{arkhe}
\title{arkhe: Tools for Cleaning Rectangular Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A dependency-free collection of simple functions for cleaning rectangular data. This package allows to detect, count and replace values or discard rows/columns using a predicate function. In addition, it provides tools to check conditions and return informative error messages.
}
\details{
\tabular{ll}{
\strong{Package:} \tab arkhe \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.1.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab \doi{10.5281/zenodo.3526659} \cr
}
}
\section{Package options}{

\code{arkhe} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{arkhe.verbose}: a \code{\link{logical}} scalar. Should \R report extra information
on progress? Defaults to \code{TRUE}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/arkhe/}
  \item \url{https://github.com/tesselle/arkhe}
  \item Report bugs at \url{https://github.com/tesselle/arkhe/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order):

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
