% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_length}
\alias{assert_length}
\alias{assert_lengths}
\alias{assert_empty}
\alias{assert_filled}
\alias{assert_dimensions}
\title{Check Object Length/Dimensions}
\usage{
assert_length(x, expected, empty = FALSE)

assert_lengths(x, expected)

assert_empty(x)

assert_filled(x)

assert_dimensions(x, expected)
}
\arguments{
\item{x}{An object to be checked.}

\item{expected}{An appropriate expected value.}

\item{empty}{A \code{\link{logical}} scalar: should empty objects be ignored?}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Object Length/Dimensions
}
\seealso{
Other validation methods: 
\code{\link{assert_constant}()},
\code{\link{assert_data}},
\code{\link{assert_lower}()},
\code{\link{assert_names}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_square}()},
\code{\link{assert_type}()},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
