% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_square}
\alias{assert_square}
\alias{assert_symmetric}
\title{Check Matrix}
\usage{
assert_square(x)

assert_symmetric(x)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be checked.}
}
\value{
Throw an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Matrix
}
\seealso{
Other checking methods: 
\code{\link{assert_constant}()},
\code{\link{assert_dim}()},
\code{\link{assert_empty}()},
\code{\link{assert_infinite}()},
\code{\link{assert_length}()},
\code{\link{assert_lower}()},
\code{\link{assert_missing}()},
\code{\link{assert_names}()},
\code{\link{assert_numeric}},
\code{\link{assert_package}()},
\code{\link{assert_type}()},
\code{\link{assert_unique}()}
}
\author{
N. Frerebeau
}
\concept{checking methods}
