% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap-method}
\alias{bootstrap,numeric-method}
\title{Bootstrap Estimation}
\usage{
bootstrap(x, ...)

\S4method{bootstrap}{numeric}(
  x,
  do,
  level = 0.95,
  type = c("student", "normal"),
  probs = c(0.25, 0.75),
  n = 1000,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{...}{Extra arguments passed to \code{do}.}

\item{do}{A \code{\link{function}} that takes \code{x} as an argument
and returns a single numeric value.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.
Must be a single number between \eqn{0} and \eqn{1}.}

\item{type}{A \code{\link{character}} string giving the type of confidence
interval to be returned. It must be one "\code{student}" (default) or
"\code{normal}". Any unambiguous substring can be given.}

\item{probs}{A \code{\link{numeric}} vector of probabilities with values in
\eqn{[0,1]} (see \code{\link[stats:quantile]{stats::quantile()}}). If \code{NULL}, quantiles are not
computed.}

\item{n}{A non-negative \code{\link{integer}} giving the number of bootstrap
replications.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values be removed
from \code{x} before the sample statistics are computed?}
}
\value{
A \code{numeric} vector with the following elements:
\describe{
\item{\code{min}}{Minimum value.}
\item{\code{mean}}{Mean value.}
\item{\code{max}}{Maximum value.}
\item{\code{lower}}{Lower bound of the confidence interval.}
\item{\code{upper}}{Upper bound of the confidence interval.}
\item{\verb{Q*}}{Sample quantile to \code{*} probability.}
}
}
\description{
Creates \eqn{n} bootstrapped replicates by resampling with replacement from
the original data and applies a function.
}
\seealso{
Other statistics: 
\code{\link{jackknife}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
