% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-attribute}
\alias{check-attribute}
\alias{check_length}
\alias{check_lengths}
\alias{check_dimension}
\alias{check_names}
\alias{check_uuid}
\title{Check Object Attributes}
\usage{
check_length(x, expected)

check_lengths(x, expected = NULL)

check_dimension(x, expected)

check_names(x, expected = NULL, margin = c(1, 2))

check_uuid(x)
}
\arguments{
\item{x}{An object to be checked.}

\item{expected}{An appropriate expected value.}
}
\value{
Throw an error, if any.
}
\description{
Check Object Attributes
}
\seealso{
Other check: \code{\link{check-graph}},
  \code{\link{check-matrix}}, \code{\link{check-missing}},
  \code{\link{check-numeric}}, \code{\link{check-type}}
}
\author{
N. Frerebeau
}
\concept{check}
\keyword{internal}
