% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-graph}
\alias{check-graph}
\alias{assert_dag}
\title{Check Graph}
\usage{
assert_dag(x)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be checked.}

\item{expected}{An appropriate expected value.}
}
\value{
Throw an error, if any.
}
\description{
Check Graph
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-matrix}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
\keyword{internal}
