% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-graph}
\alias{check-graph}
\alias{check_dag}
\title{Check Graph}
\usage{
check_dag(x)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be checked.}

\item{expected}{An appropriate expected value.}
}
\value{
Throw an error, if any.
}
\description{
Check Graph
}
\seealso{
Other check: \code{\link{check-attribute}},
  \code{\link{check-matrix}}, \code{\link{check-missing}},
  \code{\link{check-numeric}}, \code{\link{check-type}}
}
\author{
N. Frerebeau
}
\concept{check}
\keyword{internal}
