% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-matrix}
\alias{check-matrix}
\alias{check_square}
\alias{check_symmetric}
\title{Check Matrix}
\usage{
check_square(x)

check_symmetric(x)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be checked.}

\item{expected}{An appropriate expected value.}
}
\value{
Throw an error, if any.
}
\description{
Check Matrix
}
\seealso{
Other check: 
\code{\link{check-attribute}},
\code{\link{check-graph}},
\code{\link{check-missing}},
\code{\link{check-numeric}},
\code{\link{check-type}}
}
\author{
N. Frerebeau
}
\concept{check}
\keyword{internal}
