% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-matrix}
\alias{check-matrix}
\alias{assert_matrix}
\title{Check Matrix}
\usage{
assert_matrix(x, expected)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be checked.}

\item{expected}{A \code{\link{character}} string specifying the expected
value. It must be one of "\code{square}" or "\code{symmetric}".}
}
\value{
Throw an error, if any.
}
\description{
Check Matrix
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-graph}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
