% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{check-matrix}
\alias{check-matrix}
\alias{assert_symmetric}
\alias{assert_square}
\title{Check Matrix}
\usage{
assert_symmetric(x)

assert_square(x)
}
\arguments{
\item{x}{A \code{\link{matrix}} to be checked.}
}
\value{
Throw an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Matrix
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-numeric-comparison}},
\code{\link{check-numeric-trend}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
