% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-missing}
\alias{check-missing}
\alias{check_missing}
\alias{check_infinite}
\title{Check Missing Values}
\usage{
check_missing(x)

check_infinite(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
Throw an error, if any.
}
\description{
Checks if an object contains any missing (\code{NA}, \code{NaN}) or infinite
(\code{Inf}) value.
}
\seealso{
Other check: \code{\link{check-attribute}},
  \code{\link{check-graph}}, \code{\link{check-matrix}},
  \code{\link{check-numeric}}, \code{\link{check-type}}
}
\author{
N. Frerebeau
}
\concept{check}
\keyword{internal}
