% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{check-numeric}
\alias{check-numeric}
\alias{assert_count}
\alias{assert_whole}
\alias{assert_positive}
\alias{assert_negative}
\alias{assert_odd}
\alias{assert_even}
\title{Check Numeric Values}
\usage{
assert_count(x, ...)

assert_whole(x, ...)

assert_positive(x, ...)

assert_negative(x, ...)

assert_odd(x, ...)

assert_even(x, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} object to be checked.}

\item{...}{Extra parameters to be passed to internal methods.}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Numeric Values
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-matrix}},
\code{\link{check-numeric-comparison}},
\code{\link{check-numeric-trend}},
\code{\link{check-type}},
\code{\link{needs}()},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
