% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check-numeric}
\alias{check-numeric}
\alias{check_numbers}
\alias{check_constant}
\title{Check Numeric Values}
\usage{
check_numbers(x, expected = c("positive", "whole", "odd", "even"), ...)

check_constant(x)
}
\arguments{
\item{x}{A \code{\link{numeric}} object to be checked.}

\item{expected}{An appropriate expected value.}
}
\value{
Throw an error, if any.
}
\description{
Check Numeric Values
}
\seealso{
Other check: 
\code{\link{check-attribute}},
\code{\link{check-graph}},
\code{\link{check-matrix}},
\code{\link{check-missing}},
\code{\link{check-type}}
}
\author{
N. Frerebeau
}
\concept{check}
\keyword{internal}
