% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{check-type}
\alias{check-type}
\alias{assert_type}
\alias{assert_scalar}
\alias{assert_function}
\title{Check Data Types}
\usage{
assert_type(x, expected)

assert_scalar(x, expected)

assert_function(x)
}
\arguments{
\item{x}{An object to be checked.}

\item{expected}{A \code{\link{character}} string specifying the expected
type. It must be one of "\code{list}", "\code{atomic}", "\code{vector}", "\code{numeric}",
"\code{integer}", "\code{double}", "\code{character}" or "\code{logical}".}
}
\value{
Throws an error, if any, and returns \code{x} invisibly otherwise.
}
\description{
Check Data Types
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-matrix}},
\code{\link{check-numeric-comparison}},
\code{\link{check-numeric-trend}},
\code{\link{check-numeric}},
\code{\link{needs}()},
\code{\link{validate}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
