% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clean.R
\docType{methods}
\name{clean}
\alias{clean}
\alias{remove_missing}
\alias{remove_missing-method}
\alias{remove_zero}
\alias{remove_zero-method}
\alias{remove_missing,matrix-method}
\alias{remove_missing,data.frame-method}
\alias{remove_zero,matrix-method}
\alias{remove_zero,data.frame-method}
\title{Data Cleaning}
\usage{
remove_missing(object, ...)

remove_zero(object, ...)

\S4method{remove_missing}{matrix}(object, margin = 1)

\S4method{remove_missing}{data.frame}(object, margin = 1)

\S4method{remove_zero}{matrix}(object, margin = 1)

\S4method{remove_zero}{data.frame}(object, margin = 1)
}
\arguments{
\item{object}{A \code{\link{matrix}} or \code{\link{data.frame}}.}

\item{...}{Currently not used.}

\item{margin}{An \code{\link{integer}} giving the subscripts which the
cleaning will be applied over (\eqn{1} indicates rows, \eqn{2} indicates
columns).}
}
\description{
Removes row/column with missing values or zeros.
}
\author{
N. Frerebeau
}
\concept{utilities}
