% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clean.R
\docType{methods}
\name{clean}
\alias{clean}
\alias{replace_NA}
\alias{replace_NA-method}
\alias{remove_NA}
\alias{remove_NA-method}
\alias{remove_empty}
\alias{remove_empty-method}
\alias{remove_zero}
\alias{remove_zero-method}
\alias{replace_NA,matrix-method}
\alias{remove_NA,matrix-method}
\alias{remove_zero,matrix-method}
\alias{remove_empty,matrix-method}
\title{Data Cleaning}
\usage{
replace_NA(x, ...)

remove_NA(x, ...)

remove_empty(x, ...)

remove_zero(x, ...)

\S4method{replace_NA}{matrix}(x, value = 0)

\S4method{remove_NA}{matrix}(x, margin = 1, finite = TRUE)

\S4method{remove_zero}{matrix}(x, margin = 1)

\S4method{remove_empty}{matrix}(x, margin = 1)
}
\arguments{
\item{x}{A \code{\link{matrix}}, a \code{\link{data.frame}} or a \verb{*Matrix} object.}

\item{...}{Currently not used.}

\item{value}{A possible value to replace missing values of \code{x}.}

\item{margin}{An \code{\link{integer}} giving the subscript which the
cleaning will be applied over (\code{1} indicates rows, \code{2} indicates
columns).}

\item{finite}{A \code{\link{logical}} scalar: should non-\code{\link{finite}} values also be
removed?}
}
\description{
Removes empty row/column or row/column with missing values or zeros.
}
\examples{
## Create a count data matrix
A <- CountMatrix(sample(1:10, 25, TRUE), nrow = 5, ncol = 5)
A[sample(1:25, 3, FALSE)] <- 0L # Add zeros
A

## Remove row with zeros
remove_zero(A, margin = 1)

## Remove column with zeros
remove_zero(A, margin = 2)
}
\author{
N. Frerebeau
}
\concept{utilities}
