% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clean.R
\docType{methods}
\name{clean}
\alias{clean}
\alias{remove_NA}
\alias{remove_NA-method}
\alias{remove_zero}
\alias{remove_zero-method}
\alias{remove_NA,matrix-method}
\alias{remove_NA,data.frame-method}
\alias{remove_zero,matrix-method}
\alias{remove_zero,data.frame-method}
\alias{remove_zero,DataMatrix-method}
\title{Data Cleaning}
\usage{
remove_NA(x, ...)

remove_zero(x, ...)

\S4method{remove_NA}{matrix}(x, margin = 1, finite = TRUE)

\S4method{remove_NA}{data.frame}(x, margin = 1, finite = TRUE)

\S4method{remove_zero}{matrix}(x, margin = 1)

\S4method{remove_zero}{data.frame}(x, margin = 1)

\S4method{remove_zero}{DataMatrix}(x, margin = 1)
}
\arguments{
\item{x}{A \code{\link{matrix}}, \code{\link{data.frame}} or
\linkS4class{DataMatrix} object.}

\item{...}{Currently not used.}

\item{margin}{An \code{\link{integer}} giving the subscripts which the
cleaning will be applied over (\eqn{1} indicates rows, \eqn{2} indicates
columns).}

\item{finite}{A \code{\link{logical}} scalar: should non-\code{\link{finite}}
values also be removed?}
}
\description{
Removes row/column with missing values or zeros.
}
\author{
N. Frerebeau
}
\concept{utilities}
