% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex-package.R
\docType{package}
\name{codex-package}
\alias{codex-package}
\alias{_PACKAGE}
\alias{codex}
\title{codex: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
\codexr{ll}{
 \strong{Package:} \tab codex \cr
 \strong{Type:} \tab Package \cr
 \strong{Version:} \tab 0.0.0.9000 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{DOI:} \tab \href{https://doi.org/xxx}{xxx}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://codex.archaeo.science}
  \item \url{http://github.com/nfrerebeau/codex}
  \item \url{https://cran.r-project.org/package=codex}
  \item Report bugs at \url{http://github.com/nfrerebeau/codex/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\codexr{ll}{
 Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
