% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/compositions.R
\docType{methods}
\name{compositions}
\alias{compositions}
\alias{mean,AbundanceMatrix-method}
\alias{cov,AbundanceMatrix,missing-method}
\alias{cov,AbundanceMatrix-method}
\alias{cor,AbundanceMatrix,missing-method}
\alias{cor,AbundanceMatrix-method}
\alias{var,AbundanceMatrix,missing-method}
\alias{var,AbundanceMatrix-method}
\title{Compositional Data Descriptive Statistics}
\usage{
\S4method{mean}{AbundanceMatrix}(x, na.rm = FALSE)

\S4method{cov}{AbundanceMatrix,missing}(x)

\S4method{cor}{AbundanceMatrix,missing}(x)

\S4method{var}{AbundanceMatrix,missing}(x)
}
\arguments{
\item{x}{An \linkS4class{AbundanceMatrix} object.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values
(including \code{NaN}) be omitted from the calculations?}
}
\value{
A \code{\link{matrix}}.
}
\description{
Compositional Data Descriptive Statistics
}
\examples{
## Create an incidence (presence/absence) matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 20)
B <- as_abundance(A)

## Mean
mean(B)

## Variation matrix
var(B)

## Correlation matrix
cor(B)

## CLR covariance matrix
cov(B)
}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
 London: Chapman and Hall.
 DOI: \href{https://dx.doi.org/10.1007/978-94-009-4109-0}{10.1007/978-94-009-4109-0}.
}
\seealso{
Other statistics: 
\code{\link{statistics}}
}
\author{
N. Frerebeau
}
\concept{statistics}
