% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{conditions}
\alias{conditions}
\alias{throw_error}
\alias{throw_warning}
\alias{catch_message}
\alias{catch_conditions}
\title{Conditions}
\usage{
throw_error(.subclass, message, call = NULL, ...)

throw_warning(.subclass, message, call = NULL, ...)

catch_message(expr)

catch_conditions(expr)
}
\arguments{
\item{.subclass}{A \code{\link{character}} string specifying the class of
the message to be returned.}

\item{message}{A \code{\link{character}} string specifying the message to be
returned.}

\item{call}{The call.}

\item{...}{Extra arguments.}

\item{expr}{An expression to be evaluated.}
}
\description{
\itemize{
\item \code{throw_error()} stops execution of the current expression and executes an
error action.
\item \code{throw_warning()} generates a warning message.
\item \code{catch_conditions()} and \code{catch_message()} handles unusual conditions.
}
}
\author{
N. Frerebeau
}
\concept{condition}
\keyword{error}
\keyword{internal}
