% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecate.R
\docType{methods}
\name{deprecate}
\alias{deprecate}
\alias{assert_numeric}
\alias{assert_trend}
\alias{assert_matrix}
\alias{remove_empty}
\title{Deprecated Methods}
\usage{
assert_numeric(x, expected, ...)

assert_trend(x, expected, ...)

assert_matrix(x, expected)

remove_empty(x, margin = 1)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} object to be checked.}

\item{expected}{A \code{\link{character}} string specifying the expected
value (see details).}

\item{...}{Extra parameters to be passed to internal methods.}

\item{margin}{A vector giving the subscripts which the function will be
applied over (\code{1} indicates rows, \code{2} indicates columns).}
}
\description{
Deprecated Methods
}
\author{
N. Frerebeau
}
\keyword{internal}
