% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecate.R
\docType{methods}
\name{deprecate}
\alias{deprecate}
\alias{as_long}
\alias{as_long-method}
\alias{as_long,matrix-method}
\alias{confidence}
\alias{confidence-method}
\alias{confidence,numeric-method}
\alias{assert_numeric}
\alias{assert_trend}
\alias{assert_relation}
\alias{assert_matrix}
\alias{remove_empty}
\title{Deprecated Methods}
\usage{
as_long(from, ...)

\S4method{as_long}{matrix}(from, factor = FALSE, reverse = FALSE)

confidence(object, ...)

\S4method{confidence}{numeric}(object, level = 0.95, type = c("student", "normal"))

assert_numeric(x, expected, ...)

assert_trend(x, expected, ...)

assert_relation(x, y, expected, ...)

assert_matrix(x, expected)

remove_empty(x, margin = 1)
}
\arguments{
\item{from}{An object to be coerced.}

\item{...}{Extra parameters to be passed to internal methods.}

\item{factor}{A \code{\link{logical}} scalar: should character string be
coerced to \code{\link{factor}}? Default to \code{FALSE}, if \code{TRUE} the original ordering is
preserved.}

\item{reverse}{A \code{\link{logical}} scalar: should the order of factor levels be
reversed? Only used if \code{factor} is \code{TRUE}. Useful for plotting.}

\item{object}{A \code{\link{numeric}} vector.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.
Must be a single number between \eqn{0} and \eqn{1}.}

\item{type}{A \code{\link{character}} string giving the type of confidence
interval to be returned. It must be one "\code{student}" (the default) or
"\code{normal}". Any unambiguous substring can be given.}

\item{x, y}{A \code{\link{numeric}} object to be checked.}

\item{expected}{A \code{\link{character}} string specifying the expected
value (see details).}

\item{margin}{A vector giving the subscripts which the function will be
applied over (\code{1} indicates rows, \code{2} indicates columns).}
}
\description{
Deprecated Methods
}
\author{
N. Frerebeau
}
\keyword{internal}
