% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecate.R
\docType{methods}
\name{deprecate}
\alias{deprecate}
\alias{confidence}
\alias{confidence-method}
\alias{confidence,numeric-method}
\title{Deprecated Methods}
\usage{
confidence(object, ...)

\S4method{confidence}{numeric}(object, level = 0.95, type = c("student", "normal"))
}
\arguments{
\item{object}{A \code{\link{numeric}} vector.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.
Must be a single number between \eqn{0} and \eqn{1}.}

\item{type}{A \code{\link{character}} string giving the type of confidence
interval to be returned. It must be one "\code{student}" (the default) or
"\code{normal}". Any unambiguous substring can be given.}
}
\description{
Deprecated Methods
}
\author{
N. Frerebeau
}
\keyword{internal}
