% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{jackknife}
\alias{jackknife}
\alias{jackknife-method}
\alias{jackknife,numeric-method}
\title{Jackknife Estimation}
\usage{
jackknife(x, ...)

\S4method{jackknife}{numeric}(x, do, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector.}

\item{...}{Extra arguments passed to \code{do}.}

\item{do}{A \code{\link{function}} that takes \code{x} as an argument and returns a single
numeric value.}
}
\value{
A \code{numeric} vector with the following elements:
\describe{
\item{\code{mean}}{The jackknife estimate of mean.}
\item{\code{bias}}{The jackknife estimate of bias.}
\item{\code{error}}{he jackknife estimate of standard error.}
}
}
\description{
Jackknife Estimation
}
\seealso{
Other statistics: 
\code{\link{bootstrap}()}
}
\author{
N. Frerebeau
}
\concept{statistics}
