% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MatrixSummary-class}
\alias{MatrixSummary-class}
\alias{.MatrixSummary}
\title{Matrix Summary}
\description{
Matrix Summary
}
\examples{
## Create a count data matrix
X <- CountMatrix(data = sample(1:10, 45, TRUE), nrow = 15)
summary(X)

## Summary by groups
set_groups(X) <- rep(c("A", "B", "C"), each = 5)
summary(X)
}
\seealso{
Other matrix: 
\code{\link{AbundanceMatrix-class}},
\code{\link{CountMatrix-class}},
\code{\link{DataMatrix-class}},
\code{\link{IncidenceMatrix-class}},
\code{\link{OccurrenceMatrix-class}},
\code{\link{SimilarityMatrix-class}},
\code{\link{StratigraphicMatrix-class}},
\code{\link{coerce}()}
}
\author{
N. Frerebeau
}
\concept{matrix}
\keyword{internal}
