% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{get_id}
\alias{get_id-method}
\alias{set_id}
\alias{set_id-method}
\alias{set_id<-}
\alias{get_method}
\alias{get_method-method}
\alias{set_method<-}
\alias{set_method-method}
\alias{get_units}
\alias{get_units-method}
\alias{get_totals}
\alias{get_totals-method}
\alias{set_totals<-}
\alias{set_totals-method}
\alias{length,DataMatrix-method}
\alias{diag,DataMatrix-method}
\alias{dim,DataMatrix-method}
\alias{row,DataMatrix-method}
\alias{col,DataMatrix-method}
\alias{nrow,DataMatrix-method}
\alias{ncol,DataMatrix-method}
\alias{dimnames,DataMatrix-method}
\alias{dimnames<-,DataMatrix-method}
\alias{rownames,DataMatrix-method}
\alias{rownames<-,DataMatrix-method}
\alias{colnames,DataMatrix-method}
\alias{colnames<-,DataMatrix-method}
\alias{get_id,ANY-method}
\alias{get_totals,AbundanceMatrix-method}
\alias{get_method,SimilarityMatrix-method}
\alias{get_units,StratigraphicMatrix-method}
\alias{set_id,ANY-method}
\alias{set_id<-,ANY-method}
\alias{set_totals<-,AbundanceMatrix-method}
\alias{set_totals,AbundanceMatrix-method}
\alias{set_method<-,SimilarityMatrix-method}
\alias{set_method,SimilarityMatrix-method}
\title{Get or Set Parts of an Object}
\usage{
get_id(x)

set_id(x)

set_id(x) <- value

get_method(x)

set_method(x) <- value

get_units(x)

get_totals(x)

set_totals(x) <- value

\S4method{length}{DataMatrix}(x)

\S4method{diag}{DataMatrix}(x)

\S4method{dim}{DataMatrix}(x)

\S4method{row}{DataMatrix}(x, as.factor = FALSE)

\S4method{col}{DataMatrix}(x, as.factor = FALSE)

\S4method{nrow}{DataMatrix}(x)

\S4method{ncol}{DataMatrix}(x)

\S4method{dimnames}{DataMatrix}(x)

\S4method{dimnames}{DataMatrix}(x) <- value

\S4method{rownames}{DataMatrix}(x)

\S4method{rownames}{DataMatrix}(x) <- value

\S4method{colnames}{DataMatrix}(x)

\S4method{colnames}{DataMatrix}(x) <- value

\S4method{get_id}{ANY}(x)

\S4method{get_totals}{AbundanceMatrix}(x)

\S4method{get_method}{SimilarityMatrix}(x)

\S4method{get_units}{StratigraphicMatrix}(x)

\S4method{set_id}{ANY}(x)

\S4method{set_id}{ANY}(x) <- value

\S4method{set_totals}{AbundanceMatrix}(x) <- value

\S4method{set_method}{SimilarityMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{object} (see
below).}

\item{as.factor}{A \code{\link{logical}} scalar: should the value be
returned as a factor of row or column labels rather than as numbers?}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\section{Functions}{
\itemize{
\item \code{length,DataMatrix-method}: Retrieve the length of an object.

\item \code{diag,DataMatrix-method}: Extract the diagonal of a matrix.

\item \code{dim,DataMatrix-method}: Retrieve or set the dimension of an object.

\item \code{row,DataMatrix-method}: Returns a matrix of integers indicating their row number
in a matrix-like object, or a factor indicating the row labels.

\item \code{col,DataMatrix-method}: Returns a matrix of integers indicating their column
number in a matrix-like object, or a factor of column labels.

\item \code{nrow,DataMatrix-method}: Return the number of rows present in \code{x}.

\item \code{ncol,DataMatrix-method}: Return the number of columns present in \code{x}.

\item \code{dimnames,DataMatrix-method}: Retrieve the dimnames of \code{x}.

\item \code{dimnames<-,DataMatrix-method}: Set the dimnames of \code{x}.

\item \code{rownames,DataMatrix-method}: Retrieve the row names of \code{x}.

\item \code{rownames<-,DataMatrix-method}: Set the row names of \code{x}.

\item \code{colnames,DataMatrix-method}: Retrieve the column names of \code{x}.

\item \code{colnames<-,DataMatrix-method}: Set the column names of \code{x}.
}}

\examples{
## Create an incidence (presence/absence) matrix
## Data will be coerced with as.logical()
A <- IncidenceMatrix(data = sample(0:1, 100, TRUE, c(1, 1/3)), nrow = 20)
## Create a count data matrix
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20)
get_id(B)

## Access
dim(B) # Get the matrix dimensions
row(B) # Get the row indexes
col(B, as.factor = TRUE) # Get the column indexes
nrow(B) # Get the number of rows
ncol(B) # Get the number of columns
dimnames(B) # Get the dimension names
rownames(B) <- LETTERS[1:20] # Set the row names
rownames(B) # Get the rownames
colnames(B) <- letters[21:25] # Set the column names
colnames(B) # Get the column names

## Subset
B[[1, 1]] # Get the first value
B[, ] # Get all values
B[1, ] # Get the first row
B[, 1] # Get the first column
B[c("A", "B", "C"), ] # Get the first three rows
B[c("A", "B", "C"), 1, drop = FALSE] # Get the first 3 rows of the first column
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
