% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{get_method}
\alias{get_method-method}
\alias{set_method<-}
\alias{set_method-method}
\alias{get_units}
\alias{get_units-method}
\alias{get_groups}
\alias{get_groups-method}
\alias{has_groups}
\alias{has_groups-method}
\alias{set_groups<-}
\alias{set_groups-method}
\alias{get_totals}
\alias{get_totals-method}
\alias{set_totals<-}
\alias{set_totals-method}
\alias{length,DataMatrix-method}
\alias{dim,DataMatrix-method}
\alias{rownames,DataMatrix-method}
\alias{rownames<-,DataMatrix-method}
\alias{rownames<-,DataMatrix,ANY-method}
\alias{colnames,DataMatrix-method}
\alias{colnames<-,DataMatrix-method}
\alias{colnames<-,DataMatrix,ANY-method}
\alias{dimnames,DataMatrix-method}
\alias{dimnames<-,DataMatrix,list-method}
\alias{diag,DataMatrix-method}
\alias{diag<-,DataMatrix-method}
\alias{has_groups,DataMatrix-method}
\alias{get_groups,DataMatrix-method}
\alias{get_totals,AbundanceMatrix-method}
\alias{get_method,SimilarityMatrix-method}
\alias{set_groups<-,DataMatrix-method}
\alias{set_groups,DataMatrix-method}
\alias{set_totals<-,AbundanceMatrix-method}
\alias{set_totals,AbundanceMatrix-method}
\alias{set_method<-,SimilarityMatrix-method}
\alias{set_method,SimilarityMatrix-method}
\title{Get or Set Parts of an Object}
\usage{
get_method(x)

set_method(x) <- value

get_units(x)

get_groups(x)

has_groups(x)

set_groups(x) <- value

get_totals(x)

set_totals(x) <- value

\S4method{length}{DataMatrix}(x)

\S4method{dim}{DataMatrix}(x)

\S4method{rownames}{DataMatrix}(x)

\S4method{rownames}{DataMatrix}(x) <- value

\S4method{colnames}{DataMatrix}(x)

\S4method{colnames}{DataMatrix}(x) <- value

\S4method{dimnames}{DataMatrix}(x)

\S4method{dimnames}{DataMatrix,list}(x) <- value

\S4method{diag}{DataMatrix}(x)

\S4method{diag}{DataMatrix}(x) <- value

\S4method{has_groups}{DataMatrix}(x)

\S4method{get_groups}{DataMatrix}(x)

\S4method{get_totals}{AbundanceMatrix}(x)

\S4method{get_method}{SimilarityMatrix}(x)

\S4method{set_groups}{DataMatrix}(x) <- value

\S4method{set_totals}{AbundanceMatrix}(x) <- value

\S4method{set_method}{SimilarityMatrix}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set element(s) (typically a
\linkS4class{DataMatrix} object).}

\item{value}{A possible value for the element(s) of \code{object}.}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to retrieve or set parts of an object.
}
\examples{
## Create an incidence (presence/absence) matrix
## Data will be coerced with as.logical()
A <- IncidenceMatrix(data = sample(0:1, 100, TRUE, c(1, 1/3)), nrow = 20)
## Create a count data matrix
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20)

## Access
dim(B) # Get the matrix dimensions
row(B) # Get the row indexes
col(B, as.factor = TRUE) # Get the column indexes
nrow(B) # Get the number of rows
ncol(B) # Get the number of columns
dimnames(B) # Get the dimension names
rownames(B) <- LETTERS[1:20] # Set the row names
rownames(B) # Get the rownames
colnames(B) <- letters[21:25] # Set the column names
colnames(B) # Get the column names

## Subset
B[[1, 1]] # Get the first value
B[1] # Get the first value
B[, ] # Get all values
B[1, , drop = FALSE] # Get the first row
B[, 1:3] # Get the first three column
}
\seealso{
Other mutator: 
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutator}
