% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/operators.R
\docType{methods}
\name{operator}
\alias{operator}
\alias{==,CountMatrix,CountMatrix-method}
\alias{-,CountMatrix,CountMatrix-method}
\alias{-,CountMatrix,numeric-method}
\alias{+,CountMatrix,CountMatrix-method}
\alias{+,CountMatrix,numeric-method}
\alias{rowSums,DataMatrix-method}
\alias{colSums,DataMatrix-method}
\alias{rowMeans,DataMatrix-method}
\alias{colMeans,DataMatrix-method}
\title{Matrix Operations}
\usage{
\S4method{==}{CountMatrix,CountMatrix}(e1, e2)

\S4method{-}{CountMatrix,CountMatrix}(e1, e2)

\S4method{-}{CountMatrix,numeric}(e1, e2)

\S4method{+}{CountMatrix,CountMatrix}(e1, e2)

\S4method{+}{CountMatrix,numeric}(e1, e2)

\S4method{rowSums}{DataMatrix}(x, na.rm = FALSE)

\S4method{colSums}{DataMatrix}(x, na.rm = FALSE)

\S4method{rowMeans}{DataMatrix}(x, na.rm = FALSE)

\S4method{colMeans}{DataMatrix}(x, na.rm = FALSE)
}
\arguments{
\item{e1, e2}{A \linkS4class{DataMatrix} object.}

\item{x}{An object on which to perform operations.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values
(including \code{NaN}) be omitted from the calculations?}
}
\value{
A coerced object.
}
\description{
Performs operation on \code{Matrix} objects.
}
\details{
\code{rowSums} and \code{colSums} form row and column sums.

\code{rowMeans} and \code{colMeans} form row and column means.
}
\examples{
## Create a count data matrix
A1 <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Row and column sums
rowSums(A1)
colSums(A1)

## Row and column means
rowMeans(A1)
colMeans(A1)
}
\author{
N. Frerebeau
}
\concept{operator}
