% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/operators.R
\docType{methods}
\name{operator}
\alias{operator}
\alias{Arith,DataMatrix,DataMatrix-method}
\alias{Arith,DataMatrix,numeric-method}
\alias{Compare,DataMatrix,DataMatrix-method}
\alias{Compare,DataMatrix,numeric-method}
\alias{Logic,DataMatrix,DataMatrix-method}
\alias{Logic,DataMatrix,numeric-method}
\alias{Summary,DataMatrix-method}
\title{Common Operations on Matrix Objects}
\usage{
\S4method{Arith}{DataMatrix,DataMatrix}(e1, e2)

\S4method{Arith}{DataMatrix,numeric}(e1, e2)

\S4method{Compare}{DataMatrix,DataMatrix}(e1, e2)

\S4method{Compare}{DataMatrix,numeric}(e1, e2)

\S4method{Logic}{DataMatrix,DataMatrix}(e1, e2)

\S4method{Logic}{DataMatrix,numeric}(e1, e2)

\S4method{Summary}{DataMatrix}(x, na.rm = FALSE)
}
\arguments{
\item{x, e1, e2}{An object (typically a \linkS4class{DataMatrix} object).}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values
(including \code{NaN}) be omitted from the calculations?}
}
\description{
Performs common operations on \code{DataMatrix} objects.
}
\section{Group Generics}{

 \linkS4class{DataMatrix} objects have support for S4 group generic
 functionality to operate within elements across objects:
 \describe{
  \item{\code{Arith}}{"\code{+}", "\code{-}", "\code{*}", "\code{^}",
  "\code{\%\%}", "\code{\%/\%}", "\code{/}"}
  \item{\code{Compare}}{"\code{==}", "\code{>}", "\code{<}", "\code{!=}",
  "\code{<=}", "\code{>=}"}
  \item{\code{Logic}}{"\code{&}", "\code{|}"}
  \item{\code{Ops}}{"\code{Arith}", "\code{Compare}", "\code{Logic}"}
  \item{\code{Summary}}{"\code{min}", "\code{max}", "\code{range}",
  "\code{prod}", "\code{sum}", "\code{any}", "\code{all}"}
 }
}

\examples{
## Create count data matrix
A <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)
B <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Compare
A == B
A != B
A > 5
A < 5
A <= 10
A >= 10

## Arith
A + B
A * 2
A ^ 2

## Summary
max(A)
min(A)
range(A)
prod(A)
sum(A)
any(A)
all(A)

## Logic
A & B
A | B
}
\author{
N. Frerebeau
}
\concept{operator}
