% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate-trend}
\alias{predicate-trend}
\alias{is_constant}
\alias{is_increasing}
\alias{is_decreasing}
\alias{is_greater}
\alias{is_lower}
\title{Numeric Trend Predicates}
\usage{
is_constant(x, tolerance = sqrt(.Machine$double.eps), na.rm = FALSE)

is_increasing(x, na.rm = FALSE)

is_decreasing(x, na.rm = FALSE)

is_greater(x, y, strict = FALSE, na.rm = FALSE)

is_lower(x, y, strict = FALSE, na.rm = FALSE)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} object to be tested.}

\item{tolerance}{A \code{\link{numeric}} scalar giving the tolerance to check within.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values (including \code{NaN})
be omitted?}

\item{strict}{A \code{\link{logical}} scalar: should strict inequality be used?}
}
\value{
A \code{\link{logical}} scalar.
}
\description{
Check numeric objects:
\itemize{
\item \code{is_constant()} checks for equality among all elements of a vector.
\item \code{is_increasing()} and \code{is_decreasing()} check if a sequence of numbers
is monotonically increasing or decreasing, respectively.
}
}
\seealso{
Other predicates: 
\code{\link{is_scalar}},
\code{\link{predicate-attributes}},
\code{\link{predicate-data}},
\code{\link{predicate-matrix}},
\code{\link{predicate-names}},
\code{\link{predicate-numeric}},
\code{\link{predicate-type}}
}
\concept{predicates}
