% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/statistics.R
\docType{methods}
\name{statistics}
\alias{statistics}
\alias{mean,AbundanceMatrix-method}
\alias{rowMeans,DataMatrix-method}
\alias{colMeans,DataMatrix-method}
\alias{rowSums,DataMatrix-method}
\alias{colSums,DataMatrix-method}
\title{Rowwise and Columnwise Statistics}
\usage{
\S4method{mean}{AbundanceMatrix}(x, na.rm = FALSE)

\S4method{rowMeans}{DataMatrix}(x, na.rm = FALSE)

\S4method{colMeans}{DataMatrix}(x, na.rm = FALSE)

\S4method{rowSums}{DataMatrix}(x, na.rm = FALSE)

\S4method{colSums}{DataMatrix}(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \linkS4class{DataMatrix} object.}

\item{na.rm}{A \code{\link{logical}} scalar: should missing values
(including \code{NaN}) be omitted from the calculations?}
}
\description{
Computes statistics over \code{DataMatrix} margins.
}
\details{
\describe{
  \item{\code{rowMeans} and \code{colMeans}}{Form row and column means.}
  \item{\code{rowSums} and \code{colSums}}{Form row and column sums.}
 }
}
\examples{
## Create count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Row and column means
rowMeans(A)
colMeans(A)

## Row and column sums
rowSums(A)
colSums(A)

## Create relative frequencies
B <- as_abundance(A)

## Compositional mean (Aitchison 1986)
mean(B)
}
\author{
N. Frerebeau
}
\concept{statistics}
