% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{validate}
\alias{validate}
\title{Validate a Condition}
\usage{
validate(expr)
}
\arguments{
\item{expr}{An object to be evaluated.}
}
\value{
Returns \code{NULL} on success, otherwise returns the error as a string.
}
\description{
Validate a Condition
}
\seealso{
Other validation methods: 
\code{\link{check-attribute}},
\code{\link{check-data}},
\code{\link{check-matrix}},
\code{\link{check-numeric-comparison}},
\code{\link{check-numeric-trend}},
\code{\link{check-numeric}},
\code{\link{check-type}},
\code{\link{needs}()}
}
\author{
N. Frerebeau
}
\concept{validation methods}
